testenv/backend/requirements.dev.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.dev.txt requirements.dev.in
#
aniso8601==7.0.0
    # via
    #   -r requirements.base.txt
    #   graphene
asgiref==3.5.2
    # via
    #   -r requirements.base.txt
    #   django
asttokens==2.0.5
    # via stack-data
attrs==21.4.0
    # via pytest
backcall==0.2.0
    # via ipython
decorator==5.1.1
    # via ipython
django==4.0.5
    # via
    #   -r requirements.base.txt
    #   django-cors-headers
    #   django-extensions
    #   django-graphql-jwt
    #   graphene-django
django-cors-headers==3.13.0
    # via -r requirements.base.txt
django-environ==0.9.0
    # via -r requirements.base.txt
django-extensions==3.1.5
    # via -r requirements.dev.in
django-graphql-jwt==0.3.4
    # via -r requirements.base.txt
executing==0.8.3
    # via stack-data
faker==13.13.0
    # via -r requirements.dev.in
graphene==2.1.9
    # via
    #   -r requirements.base.txt
    #   django-graphql-jwt
    #   graphene-django
graphene-django==2.15.0
    # via
    #   -r requirements.base.txt
    #   django-graphql-jwt
graphql-core==2.3.2
    # via
    #   -r requirements.base.txt
    #   graphene
    #   graphene-django
    #   graphql-relay
graphql-relay==2.0.1
    # via
    #   -r requirements.base.txt
    #   graphene
iniconfig==1.1.1
    # via pytest
ipython==8.4.0
    # via -r requirements.dev.in
jedi==0.18.1
    # via
    #   ipython
    #   pudb
matplotlib-inline==0.1.3
    # via ipython
packaging==21.3
    # via
    #   pudb
    #   pytest
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pluggy==1.0.0
    # via pytest
promise==2.3
    # via
    #   -r requirements.base.txt
    #   graphene-django
    #   graphql-core
    #   graphql-relay
prompt-toolkit==3.0.29
    # via ipython
psycopg2==2.9.3
    # via -r requirements.base.txt
ptyprocess==0.7.0
    # via pexpect
pudb==2022.1.1
    # via -r requirements.dev.in
pure-eval==0.2.2
    # via stack-data
py==1.11.0
    # via pytest
pygments==2.12.0
    # via
    #   ipython
    #   pudb
pyjwt==2.4.0
    # via
    #   -r requirements.base.txt
    #   django-graphql-jwt
pyparsing==3.0.9
    # via packaging
pytest==7.1.2
    # via
    #   -r requirements.dev.in
    #   pytest-django
    #   pytest-html
    #   pytest-metadata
pytest-django==4.5.2
    # via -r requirements.dev.in
pytest-html==3.1.1
    # via -r requirements.dev.in
pytest-metadata==2.0.1
    # via pytest-html
python-dateutil==2.8.2
    # via faker
rx==1.6.1
    # via
    #   -r requirements.base.txt
    #   graphql-core
singledispatch==3.7.0
    # via
    #   -r requirements.base.txt
    #   graphene-django
six==1.16.0
    # via
    #   -r requirements.base.txt
    #   graphene
    #   graphene-django
    #   graphql-core
    #   graphql-relay
    #   promise
    #   python-dateutil
    #   singledispatch
sqlparse==0.4.2
    # via
    #   -r requirements.base.txt
    #   django
stack-data==0.3.0
    # via ipython
text-unidecode==1.3
    # via
    #   -r requirements.base.txt
    #   graphene-django
tomli==2.0.1
    # via pytest
traitlets==5.3.0
    # via
    #   ipython
    #   matplotlib-inline
urwid==2.1.2
    # via
    #   pudb
    #   urwid-readline
urwid-readline==0.13
    # via pudb
wcwidth==0.2.5
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
